iptables: add helper to add and remove rules
iptables: remove unused methods, add unit tests
